requirements.txt
# ── Web Framework & API ────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn==0.32.1
streamlit==1.42.0
httpx==0.28.1
python-dotenv==1.0.1

# ── PyTorch (install FIRST, before other deps) ───────────────────────────────
# NVIDIA GPU (CUDA 12.4):
#   pip install torch==2.5.1+cu124 torchvision==0.20.1+cu124 torchaudio==2.5.1+cu124 \
#       --index-url https://download.pytorch.org/whl/cu124
# Apple Silicon / CPU:
#   pip install torch==2.5.1

# ── Core ML dependencies ────────────────────────────────────────────────────
transformers==5.3.0
sentence-transformers==5.2.3
faiss-cpu==1.9.0.post1
openpyxl==3.1.5
numpy==2.2.6
accelerate==1.13.0
bitsandbytes==0.49.2

# ── Fine-tuning dependencies ────────────────────────────────────────────────
peft==0.10.0
trl==1.0.0
datasets==4.8.4
wandb==0.25.1

# ── Dev / testing ───────────────────────────────────────────────────────────
pytest==8.3.4
